chore: S24.03 silence clang-tidy and clang-format for atomics refactor
- Drop redundant uint32_t cast on atomic_load_explicit return (clang-tidy
  readability-redundant-casting); the slot stores _Atomic uint32_t so the
  load already returns uint32_t.
- Suppress bugprone-suspicious-include on the deliberate whitebox include
  of SolidSyslogAtomicCounter.c in its test (the rationale comment above
  the line explains the link-resolution behaviour).
- Re-flow the SOLIDSYSLOG_STATIC_ASSERT in SolidSyslogStdAtomicU32.c to
  match clang-format's preferred single-line layout.

+++ Platform/Atomics/Source/SolidSyslogStdAtomicU32.c
@@ -11,8 +11,7 @@ struct SolidSyslogAtomicU32
     _Atomic uint32_t value;
 };
 
-SOLIDSYSLOG_STATIC_ASSERT(sizeof(struct SolidSyslogAtomicU32) <= sizeof(SolidSyslogAtomicU32Storage),
-                          SolidSyslogAtomicU32Storage_too_small_for_StdAtomicU32);
+SOLIDSYSLOG_STATIC_ASSERT(sizeof(struct SolidSyslogAtomicU32) <= sizeof(SolidSyslogAtomicU32Storage), SolidSyslogAtomicU32Storage_too_small_for_StdAtomicU32);
 
 struct SolidSyslogAtomicU32* SolidSyslogAtomicU32_FromStorage(SolidSyslogAtomicU32Storage* storage)
 {
@@ -27,7 +26,7 @@ void SolidSyslogAtomicU32_Init(struct SolidSyslogAtomicU32* slot, uint32_t value
 
 uint32_t SolidSyslogAtomicU32_Load(struct SolidSyslogAtomicU32* slot)
 {
-    return (uint32_t) atomic_load_explicit(&slot->value, memory_order_relaxed);
+    return atomic_load_explicit(&slot->value, memory_order_relaxed);
 }
 
 // NOLINTNEXTLINE(bugprone-easily-swappable-parameters) — CAS shape is universal (compare_exchange convention)
